Add HTML doctype to our web ui, remove IE10-specific standard mode tag.
The use of the HTML doctype is recommended for Ext4 and Ext5, and is supported by all
post 2008 browsers.
Basically it tells the browser to interpret the HTML according to standards, and not
according to what IE6 might have thought of it.
This will also drop support for IE6 and IE7, which are anyway not supported by ext5.

The HTML doctype will make all browsers switch to their standard
compliant mode, hence we don't need anymore to toggle this specifically for IE.

Signed-off-by: Emmanuel Kasper <e.kasper@proxmox.com>