ext5migrate: remove Ext.grid.feature.Chunking hack
With ExtJS 4, we introduced an override to Ext.grid.feature.Chunking
to fix scrollings problems in the grid when having a high number
of nodes. Ext.grid.feature.Chunking has been removed from ExtJS
in version 5, so we hope either the problem is fixed on ExtJS side,
or we will have to find a different workaround.