ext6migrate: fix IPRefSelector ComboGrid
The IPrefSelector ComboGrid can have selected values which are not backed
by the component store, ie the store only contains IP aliases, but
the ComboGrid can contain an IP adress not registered as an IP alias.

In that case we should not try to update the selection in the dropdown,
as the dropdown only knows about the component store.

Implementation is similar to the syncSelection() private method of the
ExtJS4 ComboBox.