Revert "ext5migrate: do not set a custome idProperty for the KeyValue model"
This reverts commit 3150236aa5ebed6565c6eaf64126bb1195693396.

Without using an idProperty in the KeyValue model, the id of each
store item is autogenerated, and the DiffStore always recreate the whole
store content every second instead of updating the fields where values
have changed. Recreating the whole store content makes the
Node Summary View flicker, as the order of items in the store change
after each reload.

Testing shows that reversing this commit do not break anything else.