web-syslog: add date span selection
This allows the user to only show logs from a specific date span.
For that two date picker and a 'Update' button were added to the
header bar.

For the until date we assume always 23:59:59 o'clock, as witouth
a specific time journalctl assumes 00:00 which would result in an
empty log if we only wanted to look at todays log.

As the LogView is shared with others (e.g. ceph log) use an boolean
variable to set if we want to show those date pickers)

Fixes timeout when the user has a long uptime and or persistent logs