manager6: web-syslog: add date span selection
This is the ExtJS 6 version from commit
db2dd0e4390005aa551239d0a6516539b5fa200a

Same code as there.
A issue with the date picker which can be triggered by scrolling up
and then selecting a date will be fixed in ExtJS 6.0.2 (not yet
released), a workaround is in the next patch.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>