BackupEdit, cleanup: remove insideUpdate construct
insideUpdate protects for a endless loop which gets caused if the sm
selection change callback updates vmidField which then triggers the
vmidField change callback which updates the selection model again,
and the circle starts again.

As this construct is rather confusing when looking first at the code
replace it with a temporary suspend of the change event during the
vmidField update.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>